refresh token, logout 구현

+++ src/main/java/store/buzzbook/authserver/controller/AuthController.java
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import store.buzzbook.authserver.dto.AuthDTO;
 import store.buzzbook.authserver.dto.JwtResponse;
 import store.buzzbook.authserver.jwt.JwtTokenProvider;
@@ -39,9 +38,6 @@ public class AuthController {
         this.redisService = redisService;
     }
 
-    /**
-     * 토큰을 발급
-     */
     @PostMapping("/token")
     public ResponseEntity<Void> generateToken(@RequestBody AuthDTO authDTO) {
         JwtResponse response = jwtTokenProvider.generateToken(authDTO);
@@ -59,41 +55,32 @@ public class AuthController {
             @RequestHeader(value = REFRESH_HEADER, required = false) String refreshToken) {
 
         if (accessToken == null && refreshToken == null) {
-            // accessToken 과 refreshToken 둘 다 null 인 경우 처리
             log.warn("토큰 정보가 없습니다.");
             return ResponseEntity.badRequest().build();
         }
 
-        // 엑세스 토큰이 만료되었는지 확인
-        if (jwtTokenProvider.validateToken(accessToken)) {
-            // 액세스 토큰이 유효하면 바로 사용자 정보 삭제
+        accessToken = extractToken(accessToken);
+        refreshToken = extractToken(refreshToken);
+
+        if (accessToken != null && jwtTokenProvider.validateToken(accessToken)) {
             String uuid = jwtTokenProvider.getUUIDFromAccessToken(accessToken);
             redisService.removeUser(uuid);
             log.debug("엑세스 토큰으로 로그아웃 함");
-        } else if (!jwtTokenProvider.validateRefreshToken(refreshToken)) {
-            // 리프레시 토큰이 만료되지 않았을 때
+        } else if (refreshToken != null && jwtTokenProvider.validateRefreshToken(refreshToken)) {
             String userId = jwtTokenProvider.getUUIDFromRefreshToken(refreshToken);
             redisService.removeUser(userId);
-            log.debug("리프레시토큰 으로 로그아웃 함");
+            log.debug("리프레시 토큰으로 로그아웃 함");
         } else {
-            // 액세스 토큰 리프레시토큰 둘 다 만료됐을 때는 아무처리도안함
-            // 레디스에 유저정보를 저장할 때 00일 뒤 자동 삭제처리를 해놨기 때문
             log.debug("토큰이 만료됐기 때문에 자동 로그아웃 처리됨");
         }
         return ResponseEntity.ok().build();
-
-
     }
 
-    /**
-     * user info 을 반환해줌
-     */
     @GetMapping("/info")
     public ResponseEntity<Map<String, Object>> getUserInfo(
             @RequestHeader(value = TOKEN_HEADER, required = false) String accessToken,
             @RequestHeader(value = REFRESH_HEADER, required = false) String refreshToken) {
         try {
-            // refresh 토큰에 Bearer 가 없을 때
             if (isTokenPresentAndValid(refreshToken)) {
                 Map<String, Object> result = new HashMap<>();
                 result.put("message", "다시 로그인해주세요.");
@@ -101,9 +88,7 @@ public class AuthController {
             }
             refreshToken = extractToken(refreshToken);
 
-            // access token 이 만료 됐을 때
             if (isTokenPresentAndValid(accessToken)) {
-                // 토큰 재발급 시작
                 JwtResponse response = jwtTokenProvider.refreshAccessToken(refreshToken);
                 if(response != null) {
                     HttpHeaders headers = new HttpHeaders();
@@ -119,7 +104,6 @@ public class AuthController {
                     return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
                 }
             } else {
-                // access token 이 만료되지 않았을 때
                 accessToken = extractToken(accessToken);
                 Map<String, Object> userInfo = jwtTokenProvider.getUserInfoFromToken(accessToken);
                 return ResponseEntity.ok().body(userInfo);
@@ -138,6 +122,9 @@ public class AuthController {
     }
 
     private String extractToken(String token) {
-        return token.substring(BEARER_PREFIX.length());
+        if (token != null && token.startsWith(BEARER_PREFIX)) {
+            return token.substring(BEARER_PREFIX.length()).trim();
+        }
+        return token != null ? token.trim() : null;
     }
 }
